+++ src/Core/Backend.Core.DDD/Validation/ValidationPipeline.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using FluentValidation;
 using MediatR;
+using MediatR.Pipeline;
 
 namespace GoldenEye.Backend.Core.DDD.Validation
 {
@@ -11,8 +12,36 @@ namespace GoldenEye.Backend.Core.DDD.Validation
     /// </summary>
     /// <typeparam name="TRequest"></typeparam>
     /// <typeparam name="TResponse"></typeparam>
+    public class ValidationPipeline<TRequest> : IRequestPreProcessor<TRequest>
+    {
+        private readonly IEnumerable<IValidator<TRequest>> _validators;
+
+        public ValidationPipeline(IEnumerable<IValidator<TRequest>> validators)
+        {
+            _validators = validators;
+        }
+
+        public Task Process(TRequest request)
+        {
+            if (_validators?.Count() == 0)
+                return Task.CompletedTask;
+
+            var context = new ValidationContext(request);
+            var failures = _validators
+                .Select(v => v.Validate(context))
+                .SelectMany(result => result.Errors)
+                .Where(f => f != null)
+                .ToList();
+
+            if (failures.Count != 0)
+            {
+                throw new ValidationException(failures);
+            }
+            return Task.CompletedTask;
+        }
+    }
+
     public class ValidationPipeline<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
-        where TRequest : IRequest<TResponse>
     {
         private readonly IEnumerable<IValidator<TRequest>> _validators;
 
@@ -37,7 +66,6 @@ namespace GoldenEye.Backend.Core.DDD.Validation
             {
                 throw new ValidationException(failures);
             }
-
             return next();
         }
     }