+++ src/Core/Backend.Core.DDD.Tests/Changelog.md
@@ -1,1 +1,17 @@
-﻿
+﻿# v2.2.0 (18.12.2017)
+
+## Changes
+
+* Aded [tests](Validation/ValidationPipelineTests.cs) for [EventStorePipeline](../Backend.Core.DDD/Validation/ValidationPipeline.cs) **[MAJOR]**
+
+# v2.0.0
+
+## Changes
+
+* Aded [tests](Events/Store/EventStorePipelineTests.cs) for [EventStorePipeline](../Backend.Core.DDD/Events/Store/EventStorePipeline.cs) **[MAJOR]**
+
+# v1.0.0
+
+## Changes
+
+* Initial structure of project **[MAJOR]**